[7] [No-Issue] Refactoring and cleanup of Kitchen Sink (#66)
* [No-Issue] Cleaning up the menu

- Removing TBC items, commenting them out in their own section for now
- Moving Cards and Navigation to Components
- Cleaning up the styles structure
- Tidied up Kitchen Sink
- Removed buttons view model
- Removed commented code
- Moving expanded cards to Misc
- Updating spacing for accuracy (#67)
