+++ Sample/Kitchen Sink/App/KitchenSinkApp.swift
@@ -14,9 +14,7 @@ struct KitchenSinkApp: App {
                         header: HeaderTitleView(title: "Styles",
                                                 subtitle: "Base atoms used across components")
                     ) {
-                        SampleScreenView(title: "Cards", destination: Cards())
                         SampleScreenView(title: "Colours", destination: Colours())
-                        SampleScreenView(title: "Navigation (TBC)", destination: Misc())
                         SampleScreenView(title: "Spacing", destination: Spacing())
                         SampleScreenView(title: "Typography", destination: Typography())
                     }
@@ -25,16 +23,9 @@ struct KitchenSinkApp: App {
                         header: HeaderTitleView(title: "Components",
                                                 subtitle: "Complex interactable UI components")
                     ) {
-                        SampleScreenView(title: "Banners (TBC)", destination: Misc())
                         SampleScreenView(title: "Buttons", destination: Buttons())
-                        SampleScreenView(title: "Chips (TBC)", destination: Misc())
+                        SampleScreenView(title: "Cards", destination: Cards())
                         SampleScreenView(title: "Inputs", destination: Inputs())
-                        SampleScreenView(title: "Menus (TBC)", destination: Misc())
-                        SampleScreenView(title: "Notifications (TBC)", destination: Misc())
-                        SampleScreenView(title: "Progress (TBC)", destination: Misc())
-                        SampleScreenView(title: "Selection Controls (TBC)", destination: Misc())
-                        SampleScreenView(title: "Sliders (TBC)", destination: Misc())
-                        SampleScreenView(title: "Misc (To Delete)", destination: Misc())
                     }
                     .headerProminence(.increased)
                     Section(
@@ -44,6 +35,20 @@ struct KitchenSinkApp: App {
                         SampleScreenView(title: "Flight Crew Alerting", destination: Misc())
                     }
                     .headerProminence(.increased)
+//                    Section(
+//                        header: HeaderTitleView(title: "Coming Soon",
+//                                                subtitle: "Functionality coming soon to FlightUI")
+//                    ) {
+//                        SampleScreenView(title: "Banners", destination: Misc())
+//                        SampleScreenView(title: "Chips", destination: Misc())
+//                        SampleScreenView(title: "Menus", destination: Misc())
+//                        SampleScreenView(title: "Navigation", destination: Misc())
+//                        SampleScreenView(title: "Notifications", destination: Misc())
+//                        SampleScreenView(title: "Progress", destination: Misc())
+//                        SampleScreenView(title: "Selection Controls", destination: Misc())
+//                        SampleScreenView(title: "Sliders", destination: Misc())
+//                    }
+//                    .headerProminence(.increased)
                 }
                 .padding([.top], themeManager.current.padding.grid1x)
                 .scrollContentBackground(.hidden)
