--- Sample/Kitchen Sink/Buttons/Buttons+ViewModel.swift
@@ -1,7 +1,0 @@
-import SwiftUI
-import FlightUI
-
-extension Buttons {
-    class ViewModel: ObservableObject {
-    }
-}
