+++ Sample/Kitchen Sink/Components/Misc/Misc+ViewModel.swift
