+++ Sample/Kitchen Sink/Styles/Typography.swift
@@ -3,7 +3,6 @@ import FlightUI
 
 struct Typography: View {
     @EnvironmentObject var theme: Theme
-    @StateObject private var viewModel = ViewModel()
 
     var body: some View {
         ScrollView {
