--- Sample/Kitchen Sink/Typography/Typography+ViewModel.swift
@@ -1,7 +1,0 @@
-import SwiftUI
-import FlightUI
-
-extension Typography {
-    class ViewModel: ObservableObject {
-    }
-}
