+++ Sources/FlightUI/Styles/CardStyle.swift
@@ -58,7 +58,7 @@ struct CardStyleModifier: ViewModifier {
         if style.showBorder {
             RoundedRectangle(cornerRadius: style.cardRadius ?? theme.radius.medium, style: .continuous)
                 .style(
-                    withStroke: theme.color.primary.opacity(0.2),
+                    withStroke: theme.color.surfaceLow,
                     lineWidth: theme.size.border,
                     fill: style.backgroundColor ?? theme.color.surfaceLow)
         } else {
